Fix battle router
