+++ src/index.tsx
@@ -9,9 +9,9 @@ import VocabBookLayout from "./pages/VocabBook/VocabBookLayout";
 import Home from "./pages/Home/Home";
 import NoPage from "./pages/NoPage/NoPage";
 import Profile from "./pages/Profile/Profile";
-import ReviewLayout from "./pages/VocabBook/Review/Review";
+import ReviewLayout from "./pages/VocabBook/Review/ReviewLayout";
 import Review from "./pages/VocabBook/Review/Review";
-import BattleReview from "./pages/VocabBook/Review/Review";
+import BattleReview from "./pages/VocabBook/Review/BattleReview";
 import VocabBook from "./pages/VocabBook/VocabBook";
 import Wordle from "./pages/VocabBook/Wordle/Wordle";
 import { KeywordContextProvider } from "./context/keywordContext";
@@ -43,7 +43,7 @@ root.render(
                 <Route path="wordle" element={<Wordle />} />
                 <Route path="review" element={<ReviewLayout />}>
                   <Route index element={<Review />} />
-                  <Route path=":Id" element={<BattleReview />} />
+                  <Route path=":roomId" element={<BattleReview />} />
                 </Route>
               </Route>
               <Route path="profile" element={<Profile />} />
