+++ src/pages/VocabBook/Review/ReviewLayout.tsx
@@ -1,5 +1,6 @@
 import styled, { css } from "styled-components";
 import { Outlet, useOutletContext } from "react-router-dom";
+import { useViewingBook } from "../VocabBookLayout";
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 
@@ -15,7 +16,31 @@ interface Props {
 
 const Wrapper = styled.div`
   display: flex;
+  flex-direction: column;
+  align-items: center;
   padding: 20px;
+  width: 100vw;
+`;
+
+const ModeBtns = styled.div`
+  display: flex;
+  gap: 20px;
+`;
+const ReviewModeBtn = styled.button`
+  cursor: pointer;
+  ${(props: Props) =>
+    props.isBattle &&
+    css`
+      border: 1px solid gray;
+      color: gray;
+    `}
+
+  ${(props: Props) =>
+    !props.isBattle &&
+    css`
+      border: none;
+      font-weight: 600;
+    `}
 `;
 
 // type ContextType = {
@@ -38,32 +63,40 @@ const Wrapper = styled.div`
 // };
 
 export default function ReviewLayout() {
+  const navigate = useNavigate();
+  const [isBattle, setIsBattle] = useState<boolean>(false);
+  const [pin, setPin] = useState<number>();
+  const { viewingBook } = useViewingBook();
+
   useEffect(() => {
-    console.log("riewlayout");
+    console.log("riewlayout", "viewingBook", viewingBook);
+    const randomPin = Math.floor(Math.random() * 10000);
+    setPin(randomPin);
   }, []);
 
   return (
     <Wrapper>
-      <Outlet
-      // context={{
-      //   isBattle,
-      //   setIsBattle,
-      //   gameOver,
-      //   setGameOver,
-      //   round,
-      //   setRound,
-      //   pin,
-      //   setPin,
-      //   answerCount,
-      //   setAnswerCount,
-      //   reviewingQuestions,
-      //   setReviewingQuestions,
-      // }}
-      />
+      <ModeBtns>
+        <ReviewModeBtn
+        // isBattle={isBattle}
+        // onClick={() => {
+        //   setIsBattle(false);
+        //   navigate("/vocabbook/review");
+        // }}
+        >
+          Single Mode
+        </ReviewModeBtn>
+        <ReviewModeBtn
+          isBattle={!isBattle}
+          onClick={() => {
+            setIsBattle(true);
+            navigate(`/vocabbook/review/${pin}`);
+          }}
+        >
+          Battle Mode
+        </ReviewModeBtn>
+      </ModeBtns>
+      <Outlet context={{ viewingBook }} />
     </Wrapper>
   );
 }
-
-// export function useReview() {
-//   return useOutletContext<ContextType>();
-// }
